[CI] disabling make timing by default

+++ .travisCI/pdkBuild.sh
@@ -15,6 +15,10 @@
 # SPDX-License-Identifier: Apache-2.0
 export RUN_ROOT=$(pwd)
 
+
+# By default skip timing since we don't need the libs in any of the CI tests
+export SKIP_TIMING=${2:-1}
+
 docker pull efabless/openlane:staging
 
 cd $RUN_ROOT/..
@@ -25,21 +29,24 @@ echo $PDK_ROOT
 echo $RUN_ROOT
 cd $RUN_ROOT
 make skywater-pdk
-
+make skywater-library
 # The following section is for running on the CI.
 # If you're running locally you should replace them with: `make skywater-library`
 # This is because sometimes while setting up the conda env (skywater's make timing) it fails to fetch something
 # Then it exits without retrying. So, here we're retrying, and if something goes wrong it will exit after 5 retries.
 # Section Begin
-cnt=0
-until make skywater-library; do
-cnt=$((cnt+1))
-if [ $cnt -eq 5 ]; then
-	exit 2
+if [ $SKIP_TIMING -eq 0 ]; then
+	cnt=0
+	until make skywater-timing; do
+	cnt=$((cnt+1))
+	if [ $cnt -eq 5 ]; then
+		exit 2
+	fi
+	rm -rf $PDK_ROOT/skywater-pdk
+	make skywater-pdk
+	make skywater-library
+	done
 fi
-rm -rf $PDK_ROOT/skywater-pdk
-make skywater-pdk
-done
 # Section End
 
 make open_pdks
